Address code review: fix route, move types to correct namespaces
- Change PatientAccessController route from /fhir to /fhir/r4 so
  SmartScopeEnforcementMiddleware enforces JWT and scope validation.
  Extend FhirControllerBase for consistent error responses and
  auto-inject patient binding from SMART token context.

- Move IPatientAccessDataProvider and MockPatientAccessDataProvider
  from Controllers to Services namespace (FhirService.Services),
  fixing the separation of concerns issue where Program.cs had to
  reference the Controllers namespace just for DI registration.

- Move FhirClaimResource record from Mappers/PatientAccessMapper.cs
  to Models/FhirResources.cs, consolidating all lightweight FHIR
  resource definitions in one location.
